simulation_link_state_routing_protocol
The Simulation of Link-State Routing Protocol is a part of the project
under course CS542 - Computer Networks. The
program accepts the network topology details in terms of the cost of
links, and provides the shortest path tree for the
network. It uses Dijkstra’s Algorithm to construct its routing table.